[update] 整地量が小数点一桁目まで持つようにする

+++ src/main/scala/com/github/unchama/seichiassist/subsystems/breakcount/domain/level/SeichiStarLevel.scala
@@ -6,7 +6,7 @@ import com.github.unchama.seichiassist.subsystems.breakcount.domain.level.Seichi
 /**
  * 整地スターレベル。
  */
-case class SeichiStarLevel private(level: Int) extends AnyVal {
+case class SeichiStarLevel private(level: BigInt) extends AnyVal {
   def increment: SeichiStarLevel = ofNonNegative(level + 1)
 }
 
@@ -20,7 +20,7 @@ private[level] abstract class SeichiStarLevelInstances {
 
 object SeichiStarLevel extends SeichiStarLevelInstances {
 
-  def ofNonNegative(n: Int): SeichiStarLevel = {
+  def ofNonNegative(n: BigInt): SeichiStarLevel = {
     require(n >= 0, "整地スターレベルは非負である必要があります。")
     SeichiStarLevel(n)
   }
